+++ src/main/java/com/moogsan/moongsan_backend/global/infrastructure/kafka/EventPublisher.java
@@ -1,4 +1,4 @@
-package com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher;
+package com.moogsan.moongsan_backend.global.infrastructure.kafka;
 
 public interface EventPublisher {
     <T> void publish(String topic, String key, T event);
