+++ src/main/java/com/moogsan/moongsan_backend/groupbuy/application/port/BroadcastGroupBuyUpdatedEventUseCase.java
@@ -1,7 +1,7 @@
 package com.moogsan.moongsan_backend.groupbuy.application.port;
 
 import com.moogsan.moongsan_backend.groupbuy.domain.event.GroupBuyUpdatedEvent;
-import com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher.RealtimePublisher;
+import com.moogsan.moongsan_backend.groupbuy.infrastructure.kafka.RealtimePublisher;
 import com.moogsan.moongsan_backend.notification.infrastructure.template.NotificationTemplateRegistry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
