+++ src/main/java/com/moogsan/moongsan_backend/groupbuy/domain/service/GroupBuyEventService.java
@@ -3,10 +3,9 @@ package com.moogsan.moongsan_backend.groupbuy.domain.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.moogsan.moongsan_backend.domain.order.entity.Order;
-import com.moogsan.moongsan_backend.domain.order.event.OrderCanceledEvent;
 import com.moogsan.moongsan_backend.domain.order.repository.OrderRepository;
-import com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher.KafkaEventPublisher;
-import com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher.RealtimePublisher;
+import com.moogsan.moongsan_backend.global.infrastructure.kafka.KafkaEventPublisher;
+import com.moogsan.moongsan_backend.groupbuy.infrastructure.kafka.RealtimePublisher;
 import com.moogsan.moongsan_backend.groupbuy.domain.entity.GroupBuy;
 import com.moogsan.moongsan_backend.groupbuy.domain.event.GroupBuyPickupUpdatedEvent;
 import com.moogsan.moongsan_backend.groupbuy.domain.event.GroupBuyStatusEndedEvent;
