+++ src/main/java/com/moogsan/moongsan_backend/notification/domain/service/Chat/SendChatMessageUseCase.java
@@ -1,7 +1,7 @@
 package com.moogsan.moongsan_backend.notification.domain.service.Chat;
 
 import com.moogsan.moongsan_backend.groupbuy.domain.event.GroupBuyDueApproachingEvent;
-import com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher.NotificationPublisher;
+import com.moogsan.moongsan_backend.notification.infrastructure.publisher.NotificationPublisher;
 import com.moogsan.moongsan_backend.notification.infrastructure.template.NotificationTemplateRegistry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
