+++ src/main/java/com/moogsan/moongsan_backend/notification/domain/service/Order/SendOrderNotificationUseCase.java
@@ -9,7 +9,7 @@ import com.moogsan.moongsan_backend.global.infrastructure.sse.SseEmitterService;
 import com.moogsan.moongsan_backend.notification.domain.entity.NotificationType;
 import com.moogsan.moongsan_backend.notification.application.factory.NotificationFactory;
 import com.moogsan.moongsan_backend.notification.domain.repository.NotificationRepository;
-import com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher.NotificationPublisher;
+import com.moogsan.moongsan_backend.notification.infrastructure.publisher.NotificationPublisher;
 import com.moogsan.moongsan_backend.notification.infrastructure.template.NotificationTemplateRegistry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
