+++ src/main/java/com/moogsan/moongsan_backend/participantchat/domain/service/ParticipantChatEventService.java
@@ -2,8 +2,7 @@ package com.moogsan.moongsan_backend.participantchat.domain.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.moogsan.moongsan_backend.global.infrastructure.kafka.publisher.KafkaEventPublisher;
-import com.moogsan.moongsan_backend.groupbuy.domain.entity.GroupBuy;
+import com.moogsan.moongsan_backend.global.infrastructure.kafka.KafkaEventPublisher;
 import com.moogsan.moongsan_backend.participantchat.domain.entity.ChatMessageDocument;
 import com.moogsan.moongsan_backend.participantchat.domain.entity.ChatRoom;
 import com.moogsan.moongsan_backend.participantchat.domain.event.ChatMessagePersistEvent;
